requirements.txt
Flask>=1.0.0
jpype1
numpy
tensorflow
keras
